moved the common functions to a class in utils
