+++ localstack/utils/aws/aws_responses.py
@@ -3,6 +3,8 @@ import json
 from flask import Response
 from requests.models import Response as RequestsResponse
 from localstack.utils.common import to_str
+from localstack.constants import TEST_AWS_ACCOUNT_ID, MOTO_ACCOUNT_ID
+from localstack.utils.aws import aws_stack
 
 
 def flask_error_response(msg, code=500, error_type='InternalFailure'):
@@ -42,3 +44,39 @@ def requests_to_flask_response(r):
 def response_regex_replace(response, search, replace):
     response._content = re.sub(search, replace, to_str(response._content), flags=re.DOTALL | re.MULTILINE)
     response.headers['Content-Length'] = str(len(response._content))
+
+
+class MessageConversion(object):
+
+    @staticmethod
+    def _fix_date_format(response):
+        """ Normalize date to format '2019-06-13T18:10:09.1234Z' """
+
+        def _replace(response, pattern, replacement):
+            content = to_str(response.content)
+            response._content = re.sub(pattern, replacement, content)
+
+        pattern = r'<CreateDate>([^<]+) ([^<+]+)(\+[^<]*)?</CreateDate>'
+        replacement = r'<CreateDate>\1T\2Z</CreateDate>'
+        _replace(response, pattern, replacement)
+
+    @staticmethod
+    def _fix_account_id(response):
+        return aws_stack.fix_account_id_in_arns(
+            response, existing=MOTO_ACCOUNT_ID, replace=TEST_AWS_ACCOUNT_ID)
+
+    @staticmethod
+    def _fix_error_codes(method, data, response):
+        if method == 'POST' and 'Action=CreateRole' in to_str(data) and response.status_code >= 400:
+            content = to_str(response.content)
+            flags = re.MULTILINE | re.DOTALL
+            # remove the <Errors> wrapper element, as this breaks AWS Java SDKs (issue #2231)
+            response._content = re.sub(r'<Errors>\s*(<Error>(\s|.)*</Error>)\s*</Errors>', r'\1', content, flags)
+
+    @staticmethod
+    def _reset_account_id(data):
+        """ Fix account ID in request payload. All external-facing responses contain our
+            predefined account ID (defaults to 000000000000), whereas the backend endpoint
+            from moto expects a different hardcoded account ID (123456789012). """
+        return aws_stack.fix_account_id_in_arns(
+            data, colon_delimiter='%3A', existing=TEST_AWS_ACCOUNT_ID, replace=MOTO_ACCOUNT_ID)
