+++ tests/integration/test_sts.py
@@ -8,7 +8,7 @@ class TestSTSIntegrations(unittest.TestCase):
 
     def test_assume_role(self):
         test_role_session_name = 's3-access-example'
-        test_role_arn = 'arn:aws:sts::123456789012:role/rd_role'
+        test_role_arn = 'arn:aws:sts::000000000000:role/rd_role'
         response = self.sts_client.assume_role(RoleArn=test_role_arn, RoleSessionName=test_role_session_name)
 
         self.assertTrue(response['Credentials'])
